refactor: split attribution queries
