+++ src/ElectronBackend/api/queries.ts
@@ -22,6 +22,14 @@ import {
   AttributionResourceAccess,
   EDITABLE_ATTRIBUTION_RESOURCE_ACCESS,
 } from '../types/types';
+import { listAttributionRelationCounts } from './attributions/attribution-relation-counts';
+import { getAttributionSelectionSummary } from './attributions/attribution-selection-queries';
+import { getAttributions } from './attributions/getAttributions';
+import {
+  listAttributionPreview,
+  listAttributionsPage,
+  locateAttribution,
+} from './attributions/listAttributionsPage';
 import {
   getFilterExpression,
   getFilterKeys,
@@ -30,14 +38,6 @@ import {
   getSearchExpression,
   getValueFilterExpression,
 } from './filters';
-import { getAttributions } from './getAttributions';
-import {
-  getAttributionSelectionSummary,
-  listAttributionPreview,
-  listAttributionRelationCounts,
-  listAttributionsPage,
-  locateAttribution,
-} from './listAttributionsPage';
 import {
   getAttributionProgressBarData,
   getClassificationProgressBarData,
